Cleanup the user utils
The `src/utils.js` module should provide function to the user, to be
used via the console.
