+++ src/brain_stats.js
@@ -143,13 +143,26 @@ brain.stats.addRoom = function(roomName, previousCpu) {
   return true;
 };
 
+/**
+ * cpuLimit
+ * sigmoid on Game.cpu.limit + Game.cpu.bucket
+ *
+ * @return {number}
+ */
+function cpuLimit() {
+  // https://en.wikipedia.org/wiki/Sigmoid_function
+  const sigmoid = (x) => 1 + Math.tanh((2 * x) - 1);
+  return _.ceil(Game.cpu.limit * sigmoid(Game.cpu.bucket / 10000));
+}
+module.exports.cpuLimit = cpuLimit;
+
 brain.stats.updateCpuStats = function() {
   if (config.cpuStats.enabled) {
     Memory.cpuStats.last = {
       load: _.round(Game.cpu.getUsed()),
       time: Game.time,
       bucket: Game.cpu.bucket,
-      tickLimit: global.cpuLimit(),
+      tickLimit: cpuLimit(),
     };
     Memory.cpuStats.summary = {
       maxBucket: Math.max(Memory.cpuStats.summary.maxBucket, Memory.cpuStats.last.bucket),
